agent/pyproject.toml
[tool.poetry]
name = "travel"
version = "0.1.0"
description = ""
authors = ["Tyler Slaton <tyler@copilotkit.ai>"]
readme = "README.md"

[project]
name = "travel"
version = "0.1.0"
dependencies = [
    "langgraph",
    "langchain_core",
    "langchain_openai",
    "langchain-google-genai",
    "langchain",
    "openai",
    "langchain-community",
    "tavily-python",
    "python-dotenv",
    "uvicorn",
    "copilotkit==0.1.41",
    "googlemaps",
    "html2text",
    "ormsgpack (>=1.12.2,<2.0.0)",
    "langchain-groq (==0.2.5)",
    "langsmith (>=0.7.17,<0.8.0)"
]

[tool.poetry.dependencies]
python = ">=3.10.11,<3.13"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.41"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
tavily-python = "^0.5.0"
html2text = "^2024.2.26"
googlemaps = "^4.10.0"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langchain-core = "^0.3.25"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = "travel.demo:main"
